+++ src/Zicht/Itertools/util/Conversions.php
@@ -5,11 +5,9 @@
 
 namespace Zicht\Itertools\util;
 
-use Zicht\Itertools\conversions as conversion_functions;
+use Doctrine\Common\Collections\Collection;
+use Zicht\Itertools\lib\StringIterator;
 
-/**
- * @deprecated Use \Zicht\Itertools\conversions, will be removed in version 3.0
- */
 class Conversions
 {
     /**
@@ -23,48 +21,146 @@ class Conversions
      *
      * @param array|string|\Iterator $iterable
      * @return \Iterator
-     * @deprecated Use \Zicht\Itertools\conversions\mixed_to_iterator, will be removed in version 3.0
      */
     public static function mixedToIterator($iterable)
     {
-        return conversion_functions\mixed_to_iterator($iterable);
+        // NULL is often used to indicate that nothing is there,
+        // for robustness we will deal with NULL as it is an empty array
+        if (is_null($iterable)) {
+            $iterable = new \ArrayIterator([]);
+        }
+
+        // an array is *not* an instance of Traversable (as it is not an
+        // object and hence can not 'implement Traversable')
+        if (is_array($iterable)) {
+            $iterable = new \ArrayIterator($iterable);
+        }
+
+        // a string is considered iterable in Python
+        if (is_string($iterable)) {
+            $iterable = new StringIterator($iterable);
+        }
+
+        // a doctrine Collection (i.e. Array or Persistent) is also an iterator
+        if ($iterable instanceof Collection) {
+            $iterable = $iterable->getIterator();
+        }
+
+        if ($iterable instanceof \Traversable and !($iterable instanceof \Iterator)) {
+            $iterable = new \IteratorIterator($iterable);
+        }
+
+        // by now it should be an Iterator, otherwise throw an exception
+        if (!($iterable instanceof \Iterator)) {
+            throw new \InvalidArgumentException(
+                sprintf(
+                    'Argument $iterable must be a Traversable, instead %s was given',
+                    is_object($iterable) ? get_class($iterable) : gettype($iterable)
+                )
+            );
+        }
+
+        return $iterable;
     }
 
     /**
      * Try to transforms something into a Closure.
      *
-     * When $CLOSURE is null the returned Closure behaves like an identity function,
+     * When $closure is null the returned Closure behaves like an identity function,
      * i.e. it will return the value that it is given.
      *
      * @param null|\Closure $closure
      * @return \Closure
-     * @deprecated Use \Zicht\Itertools\conversions\mixed_to_closure, will be removed in version 3.0
      */
     public static function mixedToClosure($closure)
     {
-        return conversion_functions\mixed_to_closure($closure);
+        if (is_null($closure)) {
+            return function ($value) {
+                return $value;
+            };
+        }
+
+        if (!($closure instanceof \Closure)) {
+            // A \Closure is always callable, but a callable is not always a \Closure.
+            // Checking within this if statement is a slight optimization, preventing an unnecessary function wrap
+            if (is_callable($closure)) {
+                $closure = function () use ($closure) {
+                    return call_user_func_array($closure, func_get_args());
+                };
+            } else {
+                throw new \InvalidArgumentException(
+                    sprintf(
+                        'Argument $closure must be a Closure, instead %s was given',
+                        is_object($closure) ? get_class($closure) : gettype($closure)
+                    )
+                );
+            }
+        }
+
+        return $closure;
     }
 
     /**
-     * Try to transforms something into a Closure that gets a value from $STRATEGY.
+     * Try to transforms something into a Closure that gets a value from $strategy.
      *
-     * When $STRATEGY is null the returned Closure behaves like an identity function,
+     * When $strategy is null the returned Closure behaves like an identity function,
      * i.e. it will return the value that it is given.
      *
-     * When $STRATEGY is a string the returned Closure tries to find a properties,
-     * methods, or array indexes named by the string.  Multiple property, method,
-     * or index names can be separated by a dot.
-     * - 'getId'
-     * - 'getData.key'
+     * When $strategy is callable it is converted into a Closure (see mixedToClosure).
      *
-     * When $STRATEGY is callable it is converted into a Closure (see mixedToClosure).
+     * When $strategy is a string the returned Closure tries to find properties,
+     * methods, or array indexes named by the string.  Multiple property, method,
+     * or index names can be separated by a dot.  The same behavior as Twig is
+     * used, see http://twig.sensiolabs.org/doc/2.x/templates.html#variables
      *
      * @param null|string|\Closure $strategy
      * @return \Closure
-     * @deprecated Use \Zicht\Itertools\conversions\mixed_to_value_getter, will be removed in version 3.0
      */
     public static function mixedToValueGetter($strategy)
     {
-        return conversion_functions\mixed_to_value_getter($strategy);
+        if (is_string($strategy)) {
+            $keyParts = explode('.', $strategy);
+            $strategy = function ($value) use ($keyParts) {
+                foreach ($keyParts as $keyPart) {
+                    if (is_array($value) && array_key_exists($keyPart, $value)) {
+                        $value = $value[$keyPart];
+                        continue;
+                    }
+
+                    if (is_object($value)) {
+                        // property_exists does not distinguish between public, protected, or private properties, hence we need to use reflection
+                        $reflection = new \ReflectionObject($value);
+                        if ($reflection->hasProperty($keyPart)) {
+                            $property = $reflection->getProperty($keyPart);
+                            if ($property->isPublic()) {
+                                $value = $property->getValue($value);
+                                continue;
+                            }
+                        }
+
+                        foreach (['', 'get', 'is', 'has'] as $prefix) {
+                            $method = sprintf('%s%s', $prefix, $keyPart);
+                            if (is_callable([$value, $method])) {
+                                $value = call_user_func([$value, $method]);
+                                continue 2;
+                            }
+                        }
+
+                        if (method_exists($value, '__get')) {
+                            $value = $value->$keyPart;
+                            continue;
+                        }
+                    }
+
+                    // no match found
+                    $value = null;
+                    break;
+                }
+
+                return $value;
+            };
+        }
+
+        return self::mixedToClosure($strategy);
     }
 }